# api/openapi.yaml
openapi: 3.1.0
info:
  title: hv-go-ms-resume API
  description: |
    Microservicio REST para gestionar la información del portfolio personal y CV.
    Migración de Spring Boot a Go 1.22+ con arquitectura hexagonal.

    - Endpoints `/public/*` y `/login`: sin autenticación
    - Endpoints CRUD: requieren JWT Bearer Token
    - Rate limiting aplicado en todos los endpoints
  version: 2.0.0
  contact:
    name: cristiansrc
    url: https://cristiansrc.com
    email: contacto@cristiansrc.com
servers:
  - url: http://localhost:8080
    description: Desarrollo local
  - url: https://api.cristiansrc.com
    description: Producción
paths:
...
  /v1/ms-resume/public/info-page:
    get:
      operationId: getInfoPage
      summary: Obtener información completa de la página
      description: Retorna Home, BasicData, Skills (lista plana), Experiences, Educations y Altcha Challenge.
      tags:
        - Public
      security: []
      responses:
        "200":
          description: Información completa obtenida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfoPageResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/ms-resume/public/blog:
    get:
      operationId: getPublicBlogPage
      summary: Obtener artículos del blog con paginación
      description: Lista paginada de artículos del blog tecnológico.
      tags:
        - Public
      security: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Número de página (base 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Elementos por página
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: "id,desc"
          description: "Campo y dirección (ej: id,desc)"
      responses:
        "200":
          description: Lista paginada obtenida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/ms-resume/public/blog/{id}:
    get:
      operationId: getPublicBlogById
      summary: Obtener artículo del blog por ID
      tags:
        - Public
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Artículo obtenido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/ms-resume/public/blog-type:
    get:
      operationId: getPublicBlogTypes
      summary: Obtener todos los tipos de blog
      tags:
        - Public
      security: []
      responses:
        "200":
          description: Lista de tipos obtenida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogTypeResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/ms-resume/public/blog-type/{id}:
    get:
      operationId: getPublicBlogTypeById
      summary: Obtener tipo de blog por ID
      tags:
        - Public
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Tipo obtenido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogTypeResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/ms-resume/public/courses:
    get:
      operationId: getPublicCourses
      summary: Listar cursos (solo lectura)
      tags:
        - Public
      security: []
      responses:
        "200":
          description: Lista de cursos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/ms-resume/public/certifications:
    get:
      operationId: getPublicCertifications
      summary: Listar certificaciones (solo lectura)
      tags:
        - Public
      security: []
      responses:
        "200":
          description: Lista de certificaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CertificationResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/ms-resume/public/references:
    get:
      operationId: getPublicReferences
      summary: Listar referencias profesionales (solo lectura)
      tags:
        - Public
      security: []
      responses:
        "200":
          description: Lista de referencias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferenceResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/ms-resume/public/custom-sections:
    get:
      operationId: getPublicCustomSections
      summary: Listar secciones personalizadas visibles (solo lectura)
      tags:
        - Public
      security: []
      responses:
        "200":
          description: Lista de secciones custom visibles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomSectionResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ==================== BASIC DATA ====================
...
components:
  schemas:
...
    AltchaChallengeResponse:
      type: object
      properties:
        algorithm:
          type: string
          example: "SHA-256"
        challenge:
          type: string
        salt:
          type: string
        signature:
          type: string
...
    BasicDataResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        othersName:
          type: string
        firstSurName:
          type: string
        othersSurName:
          type: string
        dateBirth:
          type: string
          format: date
        located:
          type: string
        locatedEng:
          type: string
        startWorkingDate:
          type: string
          format: date
        greeting:
          type: string
        greetingEng:
          type: string
        email:
          type: string
          format: email
        instagram:
          type: string
        linkedin:
          type: string
        x:
          type: string
        github:
          type: string
        description:
          type: string
        descriptionEng:
          type: string
        descriptionPdf:
          type: array
          items:
            type: string
        descriptionPdfEng:
          type: array
          items:
            type: string
        wrapper:
          type: array
          items:
            type: string
        wrapperEng:
          type: array
          items:
            type: string
...
    ImageUrlResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameEng:
          type: string
        url:
          type: string
...
    VideoUrlResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameEng:
          type: string
        url:
          type: string
...
    LabelResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameEng:
          type: string
...
    HomeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        greeting:
          type: string
        greetingEng:
          type: string
        imageUrl:
          $ref: "#/components/schemas/ImageUrlResponse"
        buttonWorkLabel:
          type: string
        buttonWorkLabelEng:
          type: string
        buttonContactLabel:
          type: string
        buttonContactLabelEng:
          type: string
        labels:
          type: array
          items:
            $ref: "#/components/schemas/LabelResponse"
...
    BlogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        titleEng:
          type: string
        cleanUrlTitle:
          type: string
        descriptionShort:
          type: string
        description:
          type: string
        descriptionShortEng:
          type: string
        descriptionEng:
          type: string
        imageUrl:
          $ref: "#/components/schemas/ImageUrlResponse"
        videoUrl:
          $ref: "#/components/schemas/VideoUrlResponse"
        blogType:
          $ref: "#/components/schemas/BlogTypeResponse"
    BlogPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/BlogResponse"
        pageable:
          type: object
          properties:
            sort:
              type: object
              properties:
                empty:
                  type: boolean
                sorted:
                  type: boolean
                unsorted:
                  type: boolean
            offset:
              type: integer
              format: int64
            pageSize:
              type: integer
            pageNumber:
              type: integer
            paged:
              type: boolean
            unpaged:
              type: boolean
        last:
          type: boolean
        totalPages:
          type: integer
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
        number:
          type: integer
        sort:
          type: object
          properties:
            empty:
              type: boolean
            sorted:
              type: boolean
            unsorted:
              type: boolean
        first:
          type: boolean
        numberOfElements:
          type: integer
        empty:
          type: boolean
...
    BlogTypeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameEng:
          type: string
...
    SkillResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameEng:
          type: string
        skillSons:
          type: array
          items:
            $ref: "#/components/schemas/SkillSonResponse"
...
    SkillSonResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameEng:
          type: string
...
    ExperienceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        yearStart:
          type: string
          format: date
        yearEnd:
          type: string
          format: date
        company:
          type: string
        location:
          type: string
        locationEng:
          type: string
        position:
          type: string
        positionEng:
          type: string
        summary:
          type: string
        summaryEng:
          type: string
        summaryPdf:
          type: string
        summaryPdfEng:
          type: string
        descriptionItemsPdf:
          type: array
          items:
            type: string
        descriptionItemsPdfEng:
          type: array
          items:
            type: string
        skillSons:
          type: array
          items:
            $ref: "#/components/schemas/SkillSonResponse"
...
    EducationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        institution:
          type: string
        area:
          type: string
        areaEng:
          type: string
        degree:
          type: string
        degreeEng:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        location:
          type: string
        locationEng:
          type: string
        highlights:
          type: array
          items:
            type: string
        highlightsEng:
          type: array
          items:
            type: string
...
    InfoPageResponse:
      type: object
      properties:
        home:
          $ref: "#/components/schemas/HomeResponse"
        basicData:
          $ref: "#/components/schemas/BasicDataResponse"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/SkillResponse"
        experiences:
          type: array
          items:
            $ref: "#/components/schemas/ExperienceResponse"
        educations:
          type: array
          items:
            $ref: "#/components/schemas/EducationResponse"
        altchaChallenge:
          $ref: "#/components/schemas/AltchaChallengeResponse"

    # ==================== NEW ENTITY SCHEMAS ====================
...
    CourseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameEng:
          type: string
        institution:
          type: string
        institutionEng:
          type: string
        completionDate:
          type: string
          format: date
        description:
          type: string
        descriptionEng:
          type: string
        summaryPdf:
          type: string
        summaryPdfEng:
          type: string
        certificateUrl:
          type: string
...
    CertificationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameEng:
          type: string
        issuingOrganization:
          type: string
        issuingOrganizationEng:
          type: string
        issueDate:
          type: string
          format: date
        expirationDate:
          type: string
          format: date
        verificationUrl:
          type: string
        credentialId:
          type: string
        description:
          type: string
        descriptionEng:
          type: string
        summaryPdf:
          type: string
        summaryPdfEng:
          type: string
...
    ReferenceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        position:
          type: string
        company:
          type: string
        companyEng:
          type: string
        email:
          type: string
        phone:
          type: string
        relationship:
          type: string
        relationshipEng:
          type: string
...
    CustomSectionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        titleEng:
          type: string
        content:
          type: string
        contentEng:
          type: string
        summaryPdf:
          type: string
        summaryPdfEng:
          type: string
        visible:
          type: boolean

    # ==================== COMMON SCHEMAS ====================
...
    ApiErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - code
        - message
        - path
        - trace_id
        - details
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          minimum: 400
          maximum: 599
        error:
          type: string
        code:
          type: string
          description: Stable error code
        message:
          type: string
        path:
          type: string
        trace_id:
          type: string
          format: uuid
        details:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorDetail"
    ApiErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
        rejected_value: {}
  responses:
    BadRequest:
      description: Solicitud incorrecta - Error de validación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token obtenido vía POST /v1/ms-resume/login
security:
  - BearerAuth: []